requirements.txt
torch~=2.2.2
torchvision~=0.17.2
transformers~=4.48.3
pillow~=11.1.0
numpy~=1.24.0
tqdm~=4.67.1
PyYAML~=6.0.2
matplotlib~=3.10.0
torchmetrics~=1.6.1
pandas~=2.2.3
scikit-learn~=1.6.1
opencv-python~=4.11.0.86
gcsfs~=2025.3.2
tensorboard~=2.12.0